requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.9.2
pydantic-settings==2.1.0
python-multipart==0.0.6

# Core LangGraph & LangChain
langgraph==0.2.28
langchain==0.3.1  
langchain-openai==0.2.1
langchain-core==0.3.6

# OpenAI
openai==1.40.0
httpx==0.27.0

# Database
psycopg[binary]==3.2.3
pgvector==0.3.4

# Utilities
python-dotenv==1.0.1
tiktoken==0.7.0
email-validator==2.3.0

# CLI (for testing)
rich==13.9.2
